pyproject.toml
[project]
name = "wb-solarview"
version = "0.1.0"
description = "Web Service que irá receber requisições da interface mobile do SolarView e redirecionar para as APIs de Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asyncpg==0.29.0",
    "certifi==2025.8.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "fastapi==0.116.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.11.7",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.35",
    "starlette==0.47.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "uvicorn==0.35.0",
]
